Clean up move node aggregate exceptions and move them to separate folder
