+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Exception/ChildNodeAggregateIsAmbiguous.php
@@ -1,7 +1,6 @@
 <?php
-declare(strict_types=1);
 
-namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Exception;
 
 /*
  * This file is part of the Neos.ContentRepository package.
@@ -12,11 +11,10 @@ namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
  * information, please view the LICENSE file which was distributed with this
  * source code.
  */
-use Neos\Flow\Exception;
 
 /**
- * The exception to be thrown if a dimension space point is not yet occupied by a node in a node aggregate but is supposed to be
+ * The exception to be thrown if a child node aggregate was tried to be fetched but is ambiguous
  */
-class DimensionSpacePointIsNotYetOccupied extends Exception
+final class ChildNodeAggregateIsAmbiguous extends \DomainException
 {
 }
