+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Exception/NodeAggregateDisablingStrategyIsInvalid.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Exception;
 
 /*
  * This file is part of the Neos.ContentRepository package.
@@ -14,8 +14,8 @@ namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
  */
 
 /**
- * The exception to be thrown if a node aggregate does currently not cover a given dimension space point but is supposed to be
+ * The exception to be thrown if an invalid node disabling strategy was tried to be initialized
  */
-class NodeAggregateDoesCurrentlyNotCoverDimensionSpacePoint extends \DomainException
+final class NodeAggregateDisablingStrategyIsInvalid extends \DomainException
 {
 }
