+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Exception/NodeAggregateDoesCurrentlyNotCoverDimensionSpacePoint.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Exception;
 
 /*
  * This file is part of the Neos.ContentRepository package.
@@ -13,11 +13,9 @@ namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
  * source code.
  */
 
-use Neos\Flow\Exception;
-
 /**
- * The exception to be thrown if an invalid relation distribution strategy was tried to be initialized
+ * The exception to be thrown if a node aggregate does currently not cover a given dimension space point but is supposed to be
  */
-class RelationDistributionStrategyIsInvalid extends Exception
+final class NodeAggregateDoesCurrentlyNotCoverDimensionSpacePoint extends \DomainException
 {
 }
