+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/Exception/NodeTypeNotFound.php
@@ -1,6 +1,7 @@
 <?php
+declare(strict_types=1);
 
-namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Exception;
 
 /*
  * This file is part of the Neos.ContentRepository package.
@@ -13,8 +14,8 @@ namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
  */
 
 /**
- * The exception to be thrown if a root node is tried to be created with a node type which is not of type root
+ * The exception to be thrown if a given node type was not found
  */
-final class NodeTypeIsNotOfTypeRoot extends \DomainException
+final class NodeTypeNotFound extends \OutOfBoundsException
 {
 }
